Revert "dynamically calculate the max stop factor depending on the stops of measures passed in"
because this is not the only factor necessary to calculate the minimum width. Also the length of the text itself would be necessary, but the maximum stop factor might be at a completely different zoom than the longest text (i.e. longest unit+number). So, more complexity would be needed to properly make the calculation completely dynamic. That seems too much, since if a user wants to use a custom scale bar measure, he can also just set the width of the scalebar manually as a size modifier.